+++ objectbox/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## latest
 
+* Sync: support option to enable [shared global IDs](https://sync.objectbox.io/advanced/object-ids#shared-global-ids).
+
 ## 4.0.1 (2024-05-27)
 
 * Export `ObjectWithScore` and `IdWithScore` used by the new find with score `Query` methods. [#637](https://github.com/objectbox/objectbox-dart/issues/637)